backend/requirements.txt
# ──────────────────────────────────────────────
# VibeVoca Backend — Python Dependencies
# Python 3.12+  |  Pydantic v2
# ──────────────────────────────────────────────

# ── Web Framework ────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Database (Async) ────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0          # async PostgreSQL driver (runtime)
psycopg2-binary==2.9.10  # sync PostgreSQL driver (Alembic)
alembic==1.14.0

# ── Pydantic v2 & Settings ──────────────────
pydantic[email]==2.10.4
pydantic-settings==2.7.0

# ── Auth ─────────────────────────────────────
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
requests==2.31.0

# ── HTTP Client (for external APIs) ─────────
httpx==0.28.1

# ── Dev / Linting (optional, comment out for prod) ──
# pytest==8.3.4
# pytest-asyncio==0.24.0
# httpx==0.28.1  (shared with prod)
# ruff==0.8.6
